+++ src/index.js
@@ -4,7 +4,7 @@ import './js/insert-movies';
 import './js/theme';
 import './js/loader';
 import './js/team';
-import './js/modal-team';
+// import './js/modal-team';
 // import './js/modal-items';
 import './js/movie-genres';
 import './js/search-movies';
