+++ src/js/team-items.js
@@ -1,13 +1,13 @@
 export const teamItems = [
   {
-    photo: '',
+    photo: 'https://s1.hostingkartinok.com/uploads/images/2022/06/7e29ad82784f09a2d93bc33832fc3b36.jpg',
     name: 'Наша',
     family: 'команда',
     description: '',
     rol: '',
   },
   {
-    photo: 'images/team/photo-4vh.jpg',
+    photo: 'https://ca.slack-edge.com/T02UC9J08RZ-U02V3EY7XD4-110eb73de207-512',
     name: 'Вадим',
     family: 'Гудіма',
     description: 'Підключені номери сторінок (пагінація) на головній сторінці',
@@ -21,7 +21,7 @@ export const teamItems = [
     rol: 'Developer',
   },
   {
-    photo: 'images/team/photo-6ek.jpg',
+    photo: 'https://cdn.pixabay.com/photo/2019/05/14/16/43/himilayan-blue-poppy-4202825__340.jpg',
     name: 'Єлла',
     family: 'Кадряєва',
     description: 'Реалізоване модальне вікно "Наша комада"',
@@ -35,14 +35,14 @@ export const teamItems = [
     rol: 'Developer',
   },
   {
-    photo: 'images/team/photo-1uk.jpg',
+    photo: 'https://ca.slack-edge.com/T02UC9J08RZ-U02VA0E6P98-7a33fe11371b-512',
     name: 'Уляна ',
     family: 'Комаринська',
     description: 'Реалізований пошук та відображення фільмів за ключовим словом.Підключ Loader (spinner) для асинхронних запитів',
     rol: 'Developer, Scram master',
   },
   {
-    photo: 'https://files.slack.com/files-pri/T02UC9J08RZ-F03JZ134HRD/pfoto-2.jpg',
+    photo: 'https://cdn.pixabay.com/photo/2019/05/14/16/43/himilayan-blue-poppy-4202825__340.jpg',
     name: 'Сергій',
     family: 'Кіслинський',
     description: ' Реалізоване відкриття модалки з динамічно підставленою інформацією про фільм та закриття її за натисканням на клавішу ESC.',
@@ -56,7 +56,7 @@ export const teamItems = [
     rol: 'Developer',
   },
   {
-    photo: 'images/team/photo-3si.jpg',
+    photo: 'https://ca.slack-edge.com/T02UC9J08RZ-U03065NDBM1-5d1b6546a14b-512',
     name: 'Инна',
     family: 'Савченко',
     description: 'Зверстано хедер для сторінки "My Library" (мобілка, планшет, робочий стіл)',
