+++ src/partials/modal-team.html
@@ -3,13 +3,13 @@
       <div class="modal">      
           <button class="modal__button" data-modal-team-close>
               <svg class="modal__icon" >
-                  <use href="./images/sprite.svg#icon-close"></use>
+                  <use href="./images/svg/symbol-defs.svg#icon-close"></use>
               </svg>
           </button>
           <div class="team-container"></div>
       </div>    
   </div>
-
+
