+++ src/templates/templet-team.hbs
@@ -1,7 +1,7 @@
 {{#each this}}
 <div class="team__item">
   <div class="front">
-    <img class="team__image" src="{{photo}}" alt="{{name}}"/>
+    <img class="team__image" src="{{photo}}" alt="{{rol}}"/>
       <div class="team-meta">
           <p class="team__item-title">{{name}}</p>
           <p class="team__item-title">{{family}}</p>
